refactor: remove summarize/_globals.py re-export; import EXCLUDED_REGION_LABELS directly from ip_utils
