+++ src/s3_log_extraction/summarize/__init__.py
@@ -4,10 +4,8 @@ from ._generate_archive_summaries import generate_archive_summaries
 from ._generate_archive_totals import generate_archive_totals
 from ._generate_archive_totals import generate_archive_totals
 from ._generate_archive_summaries import generate_archive_summaries
-from ._globals import EXCLUDED_REGION_LABELS
 
 __all__ = [
-    "EXCLUDED_REGION_LABELS",
     "generate_archive_totals",
     "generate_summaries",
     "generate_archive_summaries",
