--- src/s3_log_extraction/summarize/_globals.py
@@ -1,3 +1,0 @@
-from ..ip_utils._globals import EXCLUDED_REGION_LABELS
-
-__all__ = ["EXCLUDED_REGION_LABELS"]
